fix: esign worker error handle for website
